requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 任务调度
celery==5.3.4
redis==5.0.1

# AI服务
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# 对象存储
boto3==1.34.0
oss2==2.18.4

# 推送服务
python-telegram-bot==20.7

# 图片处理
Pillow==10.1.0
python-magic==0.4.27

# 环境配置
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志和监控
structlog==23.2.0
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0 